+++ cmd/approve-list-svc/.gitignore
@@ -1,0 +1,1 @@
+approve-list-svc
