+++ cmd/approve-list-svc/config_example.yaml
@@ -1,0 +1,13 @@
+# host:port to listen on
+address: :6733
+
+# Optional private key for authenticated replies. Can be either Base58 encoded private key (see `signatory-tools gen-key`)
+# of PEM encoded PKCS8 key data
+private_key: edsk3YfMTdSzkFqLtmiZoFtEN6sR9jp64zin7f4jUTJiKbTLCqkkcJ
+
+# Alternatively PEM file can be stored separately
+#private_key_file: path_to_pem_file
+
+# List of allowed addresses which can be IPv4/IPv6 addresses or CIDR network addresses
+list:
+  - 192.168.88.254
