--- cmd/whitelist-svc/config_example.yaml
@@ -1,1 +1,0 @@
-private_key: edsk3YfMTdSzkFqLtmiZoFtEN6sR9jp64zin7f4jUTJiKbTLCqkkcJ
